Update docker-stack.yml
I updated the reverse_proxy service with my working image and changed the HTTPS port to HTTP port.
Also, when use this app with Docker UCP, there was a conflict in the ports as both use port 443 at the host. Therefore I changed the host post to 8002 and the container port to 81 which is an HTTP port, not HTTPS port.